gender-age-api/requirements.txt
fastapi
uvicorn
deepface==0.0.1
opencv-python==4.6.0.66
numpy==1.24.0
pillow==9.3.0
tensorflow==2.10.0
fastapi[all]  # Optional: only if you need the extra dependencies like CORS
